+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/analysis/manual/segment/SegmentDispatcher.java
@@ -38,7 +38,6 @@ public class SegmentDispatcher implements SourceDispatcher<Segment> {
         segment.setIsError(source.getIsError());
         segment.setDataBinary(source.getDataBinary());
         segment.setTimeBucket(source.getTimeBucket());
-        segment.setTagsRawData(source.getTags());
         segment.setTags(Tag.Util.toStringList(source.getTags()));
 
         RecordStreamProcessor.getInstance().in(segment);