+++ oap-server/server-storage-plugin/storage-jdbc-hikaricp-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/jdbc/h2/dao/H2AlarmQueryDAO.java
@@ -39,24 +39,20 @@ import org.apache.skywalking.oap.server.core.storage.query.IAlarmQueryDAO;
 import org.apache.skywalking.oap.server.library.client.jdbc.hikaricp.JDBCHikariCPClient;
 import org.apache.skywalking.oap.server.library.module.ModuleManager;
 import org.apache.skywalking.oap.server.library.util.CollectionUtils;
+import static org.apache.skywalking.oap.server.storage.plugin.jdbc.h2.dao.H2TableInstaller.ID_COLUMN;
 
 public class H2AlarmQueryDAO implements IAlarmQueryDAO {
 
     private JDBCHikariCPClient client;
 
     private final ModuleManager manager;
-    private final int maxSizeOfArrayColumn;
-    private final int numOfSearchValuesPerTag;
+
     private List<String> searchableTagKeys;
 
     public H2AlarmQueryDAO(JDBCHikariCPClient client,
-                           final ModuleManager manager,
-                           final int maxSizeOfArrayColumn,
-                           final int numOfSearchValuesPerTag) {
+                           final ModuleManager manager) {
         this.client = client;
         this.manager = manager;
-        this.maxSizeOfArrayColumn = maxSizeOfArrayColumn;
-        this.numOfSearchValuesPerTag = numOfSearchValuesPerTag;
     }
 
     @Override
@@ -67,13 +63,23 @@ public class H2AlarmQueryDAO implements IAlarmQueryDAO {
                     .provider()
                     .getService(ConfigService.class);
             searchableTagKeys = Arrays.asList(configService.getSearchableAlarmTags().split(Const.COMMA));
-            if (searchableTagKeys.size() > maxSizeOfArrayColumn) {
-                searchableTagKeys = searchableTagKeys.subList(0, maxSizeOfArrayColumn);
-            }
         }
         StringBuilder sql = new StringBuilder();
         List<Object> parameters = new ArrayList<>(10);
-        sql.append("from ").append(AlarmRecord.INDEX_NAME).append(" where ");
+        sql.append("from ").append(AlarmRecord.INDEX_NAME);
+        /**
+         * This is an AdditionalEntity feature, see:
+         * {@link org.apache.skywalking.oap.server.core.storage.annotation.SQLDatabase.AdditionalEntity}
+         */
+        if (!CollectionUtils.isEmpty(tags)) {
+            for (int i = 0; i < tags.size(); i++) {
+                sql.append(" inner join ").append(AlarmRecord.ADDITIONAL_TAG_TABLE).append(" ");
+                sql.append(AlarmRecord.ADDITIONAL_TAG_TABLE + i);
+                sql.append(" on ").append(AlarmRecord.INDEX_NAME).append(".").append(ID_COLUMN).append(" = ");
+                sql.append(AlarmRecord.ADDITIONAL_TAG_TABLE + i).append(".").append(ID_COLUMN);
+            }
+        }
+        sql.append(" where ");
         sql.append(" 1=1 ");
         if (Objects.nonNull(scopeId)) {
             sql.append(" and ").append(AlarmRecord.SCOPE).append(" = ?");
@@ -91,20 +97,14 @@ public class H2AlarmQueryDAO implements IAlarmQueryDAO {
             parameters.add(keyword);
         }
         if (CollectionUtils.isNotEmpty(tags)) {
-            for (final Tag tag : tags) {
-                final int foundIdx = searchableTagKeys.indexOf(tag.getKey());
+            for (int i = 0; i < tags.size(); i++) {
+                final int foundIdx = searchableTagKeys.indexOf(tags.get(i).getKey());
                 if (foundIdx > -1) {
-                    sql.append(" and (");
-                    for (int i = 0; i < numOfSearchValuesPerTag; i++) {
-                        final String physicalColumn = AlarmRecord.TAGS + "_" + (foundIdx * numOfSearchValuesPerTag + i);
-                        sql.append(physicalColumn).append(" = ? ");
-                        parameters.add(tag.toString());
-                        if (i != numOfSearchValuesPerTag - 1) {
-                            sql.append(" or ");
-                        }
-                    }
-                    sql.append(")");
+                    sql.append(" and ").append(AlarmRecord.ADDITIONAL_TAG_TABLE + i).append(".");
+                    sql.append(AlarmRecord.TAGS).append(" = ?");
+                    parameters.add(tags.get(i).toString());
                 } else {
+                    //If the tag is not searchable, but is required, then don't need to run the real query.
                     return new Alarms();
                 }
             }