+++ oap-server/server-storage-plugin/storage-jdbc-hikaricp-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/jdbc/h2/dao/H2RecordDAO.java
@@ -19,86 +19,26 @@
 package org.apache.skywalking.oap.server.storage.plugin.jdbc.h2.dao;
 
 import java.io.IOException;
-import java.util.Arrays;
-import org.apache.skywalking.oap.server.core.Const;
-import org.apache.skywalking.oap.server.core.CoreModule;
-import org.apache.skywalking.oap.server.core.UnexpectedException;
-import org.apache.skywalking.oap.server.core.alarm.AlarmRecord;
-import org.apache.skywalking.oap.server.core.analysis.manual.log.LogRecord;
-import org.apache.skywalking.oap.server.core.analysis.manual.segment.SegmentRecord;
 import org.apache.skywalking.oap.server.core.analysis.record.Record;
-import org.apache.skywalking.oap.server.core.config.ConfigService;
 import org.apache.skywalking.oap.server.core.storage.IRecordDAO;
 import org.apache.skywalking.oap.server.core.storage.model.Model;
-import org.apache.skywalking.oap.server.core.storage.type.Convert2Entity;
+import org.apache.skywalking.oap.server.core.storage.type.HashMapConverter;
 import org.apache.skywalking.oap.server.core.storage.type.StorageBuilder;
 import org.apache.skywalking.oap.server.library.client.jdbc.hikaricp.JDBCHikariCPClient;
 import org.apache.skywalking.oap.server.library.client.request.InsertRequest;
-import org.apache.skywalking.oap.server.library.module.ModuleManager;
-import org.apache.skywalking.oap.server.storage.plugin.jdbc.JDBCConverter;
 
 public class H2RecordDAO extends H2SQLExecutor implements IRecordDAO {
     private JDBCHikariCPClient h2Client;
     private StorageBuilder<Record> storageBuilder;
-    private final int maxSizeOfArrayColumn;
 
-    public H2RecordDAO(ModuleManager manager,
-                       JDBCHikariCPClient h2Client,
-                       StorageBuilder<Record> storageBuilder,
-                       final int maxSizeOfArrayColumn,
-                       final int numOfSearchableValuesPerTag) {
+    public H2RecordDAO(JDBCHikariCPClient h2Client,
+                       StorageBuilder<Record> storageBuilder) {
         this.h2Client = h2Client;
-        try {
-            if (SegmentRecord.class
-                .equals(
-                    storageBuilder.getClass().getMethod("storage2Entity", Convert2Entity.class).getReturnType()
-                )
-            ) {
-                this.maxSizeOfArrayColumn = maxSizeOfArrayColumn;
-                final ConfigService configService = manager.find(CoreModule.NAME)
-                                                           .provider()
-                                                           .getService(ConfigService.class);
-                this.storageBuilder = new H2SegmentRecordBuilder(
-                    maxSizeOfArrayColumn,
-                    numOfSearchableValuesPerTag,
-                    Arrays.asList(configService.getSearchableTracesTags().split(Const.COMMA))
-                );
-            } else if (LogRecord.class.equals(
-                storageBuilder.getClass().getMethod("storage2Entity", Convert2Entity.class).getReturnType())) {
-                this.maxSizeOfArrayColumn = maxSizeOfArrayColumn;
-                final ConfigService configService = manager.find(CoreModule.NAME)
-                                                           .provider()
-                                                           .getService(ConfigService.class);
-                this.storageBuilder = new H2LogRecordBuilder(
-                    maxSizeOfArrayColumn,
-                    numOfSearchableValuesPerTag,
-                    Arrays.asList(configService.getSearchableLogsTags()
-                                               .split(Const.COMMA))
-                );
-            } else if (AlarmRecord.class.equals(
-                storageBuilder.getClass().getMethod("storage2Entity", Convert2Entity.class).getReturnType())) {
-                this.maxSizeOfArrayColumn = maxSizeOfArrayColumn;
-                final ConfigService configService = manager.find(CoreModule.NAME)
-                                                           .provider()
-                                                           .getService(ConfigService.class);
-                this.storageBuilder = new H2AlarmRecordBuilder(
-                    maxSizeOfArrayColumn,
-                    numOfSearchableValuesPerTag,
-                    Arrays.asList(configService.getSearchableAlarmTags()
-                                               .split(Const.COMMA))
-                );
-            } else {
-                this.maxSizeOfArrayColumn = 1;
-                this.storageBuilder = storageBuilder;
-            }
-        } catch (NoSuchMethodException e) {
-            throw new UnexpectedException(
-                "Can't find the " + storageBuilder.getClass().getName() + ".storage2Entity method.");
-        }
+        this.storageBuilder = storageBuilder;
     }
 
     @Override
     public InsertRequest prepareBatchInsert(Model model, Record record) throws IOException {
-        return getInsertExecutor(model.getName(), record, storageBuilder, new JDBCConverter.RecordToStorage(), maxSizeOfArrayColumn);
+        return getInsertExecutor(model.getName(), record, storageBuilder, new HashMapConverter.ToStorage());
     }
 }