+++ oap-server/server-storage-plugin/storage-jdbc-hikaricp-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/jdbc/mysql/MySQLAlarmQueryDAO.java
@@ -24,10 +24,8 @@ import org.apache.skywalking.oap.server.storage.plugin.jdbc.h2.dao.H2AlarmQueryD
 
 public class MySQLAlarmQueryDAO extends H2AlarmQueryDAO {
 
-    public MySQLAlarmQueryDAO(final JDBCHikariCPClient client,
-                            final ModuleManager manager,
-                            final int maxSizeOfArrayColumn, final int numOfSearchValuesPerTag) {
-        super(client, manager, maxSizeOfArrayColumn, numOfSearchValuesPerTag);
+    public MySQLAlarmQueryDAO(final JDBCHikariCPClient client, final ModuleManager manager) {
+        super(client, manager);
     }
 
     @Override